Update test with naga fixes

+++ wgpu/tests/shader_layout/input_layout.wgsl
@@ -7,7 +7,7 @@ struct InputStruct {
 var<{{storage_type}}> input: InputStruct; 
 
 @group(0) @binding(1)
-var<storage, read_write> output: array<f32>;
+var<storage, read_write> output: array<u32>;
 
 @compute @workgroup_size(1)
 fn cs_main() {
